Refactor more code
